I changed headers generation to remove memset and strcat