@jkuri 2018-07-03
This actually improves performance and throughput for a bit 👍 I made benchmarks here:

before:
```
 ~/ wrk -t20 -c1000 -d10s http://whoami.local.net
Running 10s test @ http://whoami.local.net
  20 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   137.63ms  113.72ms   1.28s    81.15%
    Req/Sec   181.14     75.37   555.00     72.57%
  35563 requests in 10.04s, 11.97MB read
  Socket errors: connect 0, read 996, write 0, timeout 0
Requests/sec:   3541.87
Transfer/sec:      1.19MB
```

after:
```
 ~/ wrk -t20 -c1000 -d10s http://whoami.local.net
Running 10s test @ http://whoami.local.net
  20 threads and 1000 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   148.83ms  140.48ms   1.76s    85.54%
    Req/Sec   188.19     98.22   585.00     72.09%
  37273 requests in 10.10s, 12.55MB read
  Socket errors: connect 0, read 977, write 0, timeout 0
Requests/sec:   3689.88
Transfer/sec:      1.24MB
```